+++ RELEASES.md
@@ -6,6 +6,7 @@
 
 * Binaries 'powercap-{info,set}-intel-rapl' (more faithful to Linux power capping framework interface than 'rapl-{info,set}')
 * Top-level 'powercap_control_type' type to powercap.h with file getters/setters (for 'enabled' field)
+* Functions 'powercap_sysfs_control_type_{get,set}_enabled' in powercap-sysfs.h
 
 ### Changed